+++ Headers/VLBCreateItemOperationDelegate.h
@@ -15,7 +15,7 @@
 
 -(void)didCancel;
 -(void)didStartUploadingItem:(UIImage*)itemImage key:(NSString*)key location:(NSDictionary*) location locality:(NSString*) locality;
--(void)bytesWritten:(NSInteger)bytesWritten totalBytesWritten:(long long)totalBytesWritten totalBytesExpectedToWrite:(long long)totalBytesExpectedToWrite;
+-(void)bytesWritten:(long long)bytesWritten totalBytesWritten:(long long)totalBytesWritten totalBytesExpectedToWrite:(long long)totalBytesExpectedToWrite;
 -(void)didSucceedWithItem:(NSDictionary*)item;
 -(void)didFailOnItemWithError:(NSError*)error;
